Fix compile warning in math/gauss.

+++ math/gauss.cc
@@ -7,12 +7,12 @@ namespace math {
 
 bool gauss(std::vector<std::vector<double>> A, std::vector<double> b, std::vector<double>* x)
 {
-    int H = A.size();
+    size_t H = A.size();
 
-    for (int i = 0; i < H; ++i) {
+    for (size_t i = 0; i < H; ++i) {
         // --- choose pivot.
-        int pivot = i;
-        for (int j = i + 1; j < H; ++j) {
+        size_t pivot = i;
+        for (size_t j = i + 1; j < H; ++j) {
             if (abs(A[pivot][i]) < abs(A[j][i]))
                 pivot = j;
         }
@@ -24,18 +24,18 @@ bool gauss(std::vector<std::vector<double>> A, std::vector<double> b, std::vecto
         // --- divide each row by A[i][i].
         {
             double v = A[i][i];
-            for (int j = i; j < A[i].size(); ++j) {
+            for (size_t j = i; j < A[i].size(); ++j) {
                 A[i][j] /= v;
             }
             b[i] /= v;
         }
 
         // --- subtract
-        for (int j = 0; j < H; ++j) {
+        for (size_t j = 0; j < H; ++j) {
             if (i == j)
                 continue;
             double v = A[j][i];
-            for (int k = i; k < A[j].size(); ++k) {
+            for (size_t k = i; k < A[j].size(); ++k) {
                 A[j][k] -= A[i][k] * v;
             }
             b[j] -= b[i] * v;
